requirements.txt
fastapi==0.115.4
numpy==2.1.3
pandas==2.2.3
Pillow==11.0.0
plotly==5.24.1
prometheus_client==0.21.0
Requests==2.32.3
streamlit==1.40.1
torch==2.5.1
transformers==4.46.2
uvicorn==0.32.0
python-multipart
